Fixed Parser bug #126
* Parser was calling trim() which advances the mParserPos after
checking hasNext() in the while conditional. In the case where the
command string left to parse is entirely whitespace, the mParserPos
will be advanced to the end of the command string. The next peek() call
in the GENERAL parser state will cause an IndexOutOfBoundsException as
peek() does not do bounds checking. The trim command has been inlined
into a switch case within the GENERAL state instead.
* AddHandler adds the newly created Task object regardless of whether
InvalidTaskException was thrown, bug has been fixed.
* Closes # 126
* Added regression test case to Parser tests for #126